+++ app/src/main/java/protect/card_locker/ImportExportTask.java
@@ -11,7 +11,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
-import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 
 import protect.card_locker.importexport.DataFormat;