+++ app/src/main/java/protect/card_locker/LoyaltyCard.java
@@ -7,8 +7,6 @@ import com.google.zxing.BarcodeFormat;
 import java.math.BigDecimal;
 import java.util.Currency;
 import java.util.Date;
-import java.util.HashMap;
-import java.util.Map;
 
 import androidx.annotation.Nullable;