+++ app/src/main/java/protect/card_locker/importexport/MultiFormatExporter.java
@@ -5,7 +5,6 @@ import android.util.Log;
 
 import java.io.IOException;
 import java.io.OutputStream;
-import java.io.OutputStreamWriter;
 
 import protect.card_locker.DBHelper;